requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-silk>=5.0.0
redis[hiredis]
django-redis>=5.3.0
